Amanding HW MVVM
