+++ app/src/main/java/ru/netology/myapp/PostRepository.kt
@@ -5,4 +5,5 @@ import androidx.lifecycle.LiveData
 interface PostRepository {
     fun get(): LiveData<Post>
     fun like()
+    fun share()
 }
