+++ app/src/main/java/ru/netology/myapp/PostViewModel.kt
@@ -6,4 +6,5 @@ class PostViewModel: ViewModel() {
     private val repository: PostRepository = PostRepositoryImpl()
     val data = repository.get()
     fun clickedLike() = repository.like()
+    fun clickedShare() = repository.share()
 }
